Include example on how to use this with Mocha/Se
Included an example on how to do this with Mocha/Se WebDriver JS. This can save someone much time, as it showcases the need for a done() callback in the test.before hook. This is necessary to work with the async nature of this browserstack-local module, otherwise a race condition is present between the local binary being run and the webdriver attempting to connect to it.